% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc funs.R
\name{binarize.vals}
\alias{binarize.vals}
\title{Convert other formats to binary flags.}
\usage{
binarize.vals(x, val0, val1, ignore.case = TRUE)
}
\arguments{
\item{x}{a vector.}

\item{val0, val1}{the values to convert to 0/1 respectively.}

\item{ignore.case}{if TRUE, conversion will ignore}

\item{as.numeric}{if TRUE, return a numeric vector else return a string.}
}
\value{
A new vector of zeroes and ones.
}
\description{
Convert other formats to binary flags.
}
\section{Description}{

Flags/indicators are stored in different formats. This function converts two arbitrary values to 0/1. Others will be converted to NA.
The default behavior ignores case, e.g. binarize.vals(x, 'Y', 'N') will convert 'Y' to 0 but 'y' to NA. Setting ignore.case = FALSE will first convert all
}

\section{Warning}{

TRUE is not 'true'! Setting ignore.case = FALSE when converting T/F will not work as desired.
The function should catch this by checking is.logical(x).
}

\examples{
x <- c(sample(c('Y', 'N'), 5, replace = T), 'a', 'y', 0, 'n')
print(x)
binarize.vals(x, 'Y', 'N')

x <- c(T, F, T, T, F)
binarize.vals(x, 'T', 'F', ignore.case = F)  # wrong

}
