% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup packages.R
\name{bulk.reup}
\alias{bulk.reup}
\title{Bulk re-install/update}
\value{
NA, quietly updates packages
}
\description{
Bulk re-install (or update) installed packages (should rarely be necessary)
}
\examples{
bulk.reinstall()
}
