% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt funs.R
\name{Cs}
\alias{Cs}
\title{Create quoted character vectors}
\usage{
Cs(...)
}
\arguments{
\item{...}{text to be concatenated.}
}
\value{
A quoted character vector.
}
\description{
Create quoted character vectors
}
\examples{
Cs(some, pig)
}
