% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt funs.R
\name{fix.sdb.names}
\alias{fix.sdb.names}
\title{Make friendlier names for SDB fields.}
\usage{
fix.sdb.names(x = "EDWSampleVarName")
}
\arguments{
\item{x}{A vector of strings, typically the result of names(...).}
}
\value{
A character vector in r-ish format, i.e. lowercase text with words separated by '.'.
}
\description{
Standardize columns names into preferred format (e.g. 'r-ify them with . as the separator)
}
\examples{
fix.sdb.names(c("SDBSrcSysKey", "QuarterlyGPA", "AcademicYear"))

}
