% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yrq funs.R
\name{qtr.diff}
\alias{qtr.diff}
\title{Difference two year-quarters}
\usage{
qtr.diff(yrq1, yrq2)
}
\arguments{
\item{yrq1, yrq2}{two 5-digit year-quarters.}
}
\value{
The difference between two academic year-quarters, expressed in quarters.
  Does not return absolute value, so if yrq1 < yrq2, result will be negative.
}
\description{
Difference two year-quarters
}
\examples{
qtr.diff(20054, 20044)
qtr.diff(19994, 20001)

x <- c(20031, 19804, 20181)
y <- c(20203, 19794, 20174)
qtr.diff(x, y)
}
