% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yrq funs.R
\name{qtrfuns}
\alias{qtrfuns}
\alias{next.qtr}
\alias{last.qtr}
\title{Next/last academic quarters}
\usage{
next.qtr(x)

last.qtr(x)
}
\arguments{
\item{x}{a quarter, numeric from 1:4 (Winter > Spring > Summer > Fall)}
}
\value{
The next, or prior, academic quarter, an integer from 1:4.
}
\description{
Simple functions to correctly increment academic quarters. I.e. 4 (Fall) should roll to 1 (Winter).
}
\examples{
next.qtr(4)
x <- c(1, 3, 4, 2, 4)
next.qtr(x)
last.qtr(x)
}
