% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc funs.R
\name{rm.empty}
\alias{rm.empty}
\title{Many tables are unused (undocumented) or constant. Remove them.}
\usage{
rm.empty(df)
}
\arguments{
\item{df}{a dataframe.}
}
\value{
A dataframe w/o zero variance or constant columns.
}
\description{
Many tables are unused (undocumented) or constant. Remove them.
}
