% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt funs.R
\name{trim.white}
\alias{trim.white}
\title{Trim excess whitespace from data frame cols}
\usage{
trim.white(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A data.frame with whitespace trimmed from characters columns.
}
\description{
Trim excess whitespace from data frame cols
}
\examples{
x <- data.frame("a" = c(" Wilbur  ", "Charlotte A. Cavatica  ", "  Fern Arable   ", " Templeton  "), "b" = 1:4)
trim.white(x)
}
