% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc funs.R
\name{varwhich}
\alias{varwhich}
\title{Return column index}
\usage{
varwhich(df, x)
}
\arguments{
\item{df}{A data frame}

\item{x}{A (quoted) column name}
}
\description{
Return column index
}
\section{Description}{

Sometimes it is useful to find or pass columns by position. This is a simple shortcut that returns the numeric index of a data frame column (or columns).
Calls internal col.loc() function (the original 1-variable version of varwhich)
}

